+++ src/main/java/dunn/ted/java/instateam/service/ProjectService.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface ProjectService {
 
     List<Project> findAll();
-    Project findById(Long id);
+    Project findById(int id);
     void save(Project project);
 }
